Initialise counters w/o using CounterVec#Set.
`CounterVec#Set` was present in `v0.8.0` but has been removed in `v0.9.0-pre1`.
Simply calling the counter with a label initialises it, so that is what we are now doing.
If not calling it at all, then the metric is not exposed for this label until actually incremented, which is probably not the desired behaviour -- i.e. people would most likely still want to know that the number of debug/info/warning/error log messages is zero, and to confirm that their monitoring is not broken.